proj2/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0
timm>=0.6.0
pandas
numpy
scikit-learn
Pillow
tqdm
nltk
pycocoevalcap
radgraph
chexpert-labeler
